new new

+++ numpy7.py
@@ -1,6 +1,6 @@
 import numpy as n
 
-arr = n.array([1, 2, 3, 4, 5], dtype = "i8")
+arr = n.array([1, 2, 3, 4, 5], dtype = "i4")
 print(arr)
 print(arr.dtype)
